Add living workflow documentation
This document is meant to serve as a living (i.e., non-static)
description of the workflow uses in this project. This initial creation
of the document describes the old workflow and the new proposed
workflow. Once the proposed workflow has been refined and agreed upon,
the document will change to only reflect the agreed upon workflow.